fix: issues with foreign key consistency accross different targets
