+++ convention/publishing/src/main/kotlin/com/attafitamim/kabin/publish/PublishConventions.kt
@@ -10,7 +10,7 @@ import org.gradle.api.publish.maven.MavenPomScm
 
 class PublishConventions : Plugin<Project> {
 
-  private val version = "0.1.0-alpha07"
+  private val version = "0.1.0-alpha08"
   private val group = "com.attafitamim.kabin"
 
   override fun apply(project: Project) {
