--- library/core/src/commonMain/kotlin/com/attafitamim/kabin/core/driver/DriverFactory.kt
@@ -1,10 +1,0 @@
-package com.attafitamim.kabin.core.driver
-
-import app.cash.sqldelight.db.QueryResult
-import app.cash.sqldelight.db.SqlDriver
-import app.cash.sqldelight.db.SqlSchema
-import com.attafitamim.kabin.core.database.configuration.KabinDatabaseConfiguration
-
-expect fun KabinDatabaseConfiguration.createDriver(
-    schema: SqlSchema<QueryResult.AsyncValue<Unit>>,
-): SqlDriver
