--- library/core/src/jsMain/kotlin/com/attafitamim/kabin/core/database/configuration/KabinDatabaseConfiguration.kt
@@ -1,5 +1,0 @@
-package com.attafitamim.kabin.core.database.configuration
-
-actual class KabinDatabaseConfiguration(
-    actual val extendedConfig: KabinExtendedConfig = KabinExtendedConfig()
-)
